Update WP Rollback plugin version to 2.0.7
This commit updates the WP Rollback plugin version to 2.0.7 and indicates compatibility with WordPress version 6.6. It includes changes to the plugin's description, along with the addition of bug fixes in the changelog. These fixes resolve an error with plain permalink websites and update the way the React app is loaded to suppress React 18+ warnings.